feat: move data block APIs to the data block module
