+++ apps/web/core/blocks/data/collection.ts
@@ -5,30 +5,6 @@ import { StoreRelation } from '~/core/database/types';
 import { DB } from '~/core/database/write';
 import { EntityId, SpaceId } from '~/core/io/schema';
 
-import { getRelationForBlockType } from './block-types';
-import { makeRelationForSourceType } from './sources';
-
-/**
- * Returns the relations to create a data entity. Data entities require a type,
- * source type, and source relations by default to be valid.
- *
- * This function returns all the relations needed to make a data entity, defaulting
- * to a source type of Collection. The initial source points to a new collection
- * created for the new data entity.
- *
- * @param blockId the id of the new data block as an {@link EntityId}
- * @returns an array of {@link StoreRelation} representing the data entity relations.
- */
-export function makeInitialDataEntityRelations(blockId: EntityId, spaceId: string): [StoreRelation, StoreRelation] {
-  return [
-    // Create relation for the source type, e.g., Spaces, Collection, Geo, etc.
-    makeRelationForSourceType('COLLECTION', blockId, spaceId),
-
-    // Create the type relation for the block itself. e.g., Table, Image, Text, etc.
-    getRelationForBlockType(blockId, SYSTEM_IDS.DATA_BLOCK, spaceId),
-  ];
-}
-
 type CreateCollectionItemRelationArgs = {
   relationId?: EntityId;
   collectionId: EntityId;
