+++ apps/web/partials/blocks/table/table-block-filter-creation-prompt.tsx
@@ -5,11 +5,11 @@ import { AnimatePresence, motion } from 'framer-motion';
 import * as React from 'react';
 
 import { Filter } from '~/core/blocks/data/filters';
+import { Source } from '~/core/blocks/data/source';
 import { useDebouncedValue } from '~/core/hooks/use-debounced-value';
 import { useSearch } from '~/core/hooks/use-search';
 import { useSpaces } from '~/core/hooks/use-spaces';
 import { Space } from '~/core/io/dto/spaces';
-import { Source } from '~/core/state/editor/types';
 import { useTableBlock } from '~/core/state/table-block-store';
 import { FilterableValueType } from '~/core/value-types';
 
@@ -160,7 +160,7 @@ const reducer = (state: PromptState, action: PromptAction): PromptState => {
 };
 
 function getInitialState(source: Source): PromptState {
-  if (source.type === 'ENTITY') {
+  if (source.type === 'RELATIONS') {
     return {
       selectedColumn: SYSTEM_IDS.RELATION_TYPE_ATTRIBUTE,
       value: {
@@ -183,7 +183,7 @@ function getInitialState(source: Source): PromptState {
 }
 
 export function TableBlockFilterPrompt({ trigger, onCreate, options }: TableBlockFilterPromptProps) {
-  const { columnRelationTypes, source } = useTableBlock();
+  const { source } = useTableBlock();
 
   const [state, dispatch] = React.useReducer(reducer, getInitialState(source));
